Adjusted cards to be dynamic not static
I originally planned to create the cards and then just activate them when necessary, but Michael's approach in the tutorial is to recreate the UI info from the Player Stats for each character in the hierarchy and use the UI elements only as templates to be filled in. So, I've adjusted my original plan to match this approach.