Quest rewards via an abstract class
Abstract class with one method, Reward(). QuestRewards class passed as an argument. Enums for PlayerClass and RewardType added to QuestRewards. QuestRewards serialized inside Quest. So, rewards can now be set inside the Inspector, with players receiving the reward and type of reward via a dropdown. Enables specialised rewards to be added quite easily, with the enum allowing for the reward type to be targeted to the playerType's override method.