Save system - replaced Save system
In the previous save system, I'd tried to copy something from a farming RPG, but it just wasn't built for my type of game. So, I've rebuilt the ISaveable interface and SaveData class from the ground up.
Tested successfully with Struct for Thulgran's data, a global LoadSave call via the GameManager with OnEnable and OnDestroy.